+++ tests/app/test_flaskutils.py
@@ -1,6 +1,6 @@
 import os
 
-import toml
+import tomlkit
 
 from canaille import create_app
 from canaille.app.flask import set_parameter_in_url_query
@@ -26,7 +26,7 @@ def test_set_parameter_in_url_query():
 def test_environment_configuration(configuration, tmp_path):
     config_path = os.path.join(tmp_path, "config.toml")
     with open(config_path, "w") as fd:
-        toml.dump(configuration, fd)
+        tomlkit.dump(configuration, fd)
 
     os.environ["CONFIG"] = config_path
     app = create_app()